added call engine response

+++ shared_model/interfaces/common_objects/types.hpp
@@ -96,7 +96,7 @@ namespace shared_model {
       /// Type of the smart contract code
       using SmartContractCodeType = std::string;
       /// Type of command index within a transaction
-      using CommandIndexType = uint64_t;
+      using CommandIndexType = int32_t;
       /// Transaction index type
       using TxIndexType = int32_t;
